refactor: move webhook package to pkg
